services/homeassistant-service/requirements.txt
flask>=2.3.3,<3.0.0
requests==2.31.0
homeassistant-api==5.0.0
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.4.2
redis==5.0.1
psycopg2-binary==2.9.9
websockets==11.0.3
retrying==1.3.4
prometheus-client==0.17.1